### 2019-03-19 db68739
Minor editorial fixes to the JOSS paper text
Fixes some typos, missing prepositions/articles, minor wording changes, etc.
