+++ paper/paper.md
@@ -30,12 +30,12 @@ This model gives a description of the interaction between the ocean waves and th
 Assuming an incompressible, inviscid and irrotational flow, the fluid motion can be modeled by a scalar field $\phi$, called velocity potential and such that $u = \nabla \phi$.
 The conservation of the mass of the fluid can be written as a Laplace problem $\nabla^2 \phi = 0$ .
 Assuming small motions of the floating body and small wave heights, the boundary condition at the free surface can be linearized.
-The resulting problem is fully linear and can be solved in frequency domain.
+The resulting problem is fully linear and can be solved in the frequency domain.
 
 The problem can be formulated as a boundary integral problem and solved by the Boundary Element Method (BEM).
-In contrast with other BEM problems, the linear potential flow solvers use an expression of the Green function which accounts for the boundary conditions at the free surface and the sea bottom.
-Thus only the wet surface of the floating body needs to be discretized.
-The Green function can be computed in several ways [@xie_comparison_2018], which are implemented in several software, most of them commercial and closed-source. 
+In contrast with other BEM problems, the linear potential flow solvers use an expression of the Green's function which accounts for the boundary conditions at the free surface and the sea bottom.
+Thus, only the wet surface of the floating body needs to be discretized.
+The Green's function can be computed in several ways [@xie_comparison_2018], which are implemented in several software, most of them commercial and closed-source. 
 
 In 2014, the École Centrale de Nantes (ECN) released the code *Nemoh* as an open-source software [@babarit_theoretical_2015].
 This solver is based on in-house and commercial codes developed in the ECN since the 1970's [@delhommeau_problemes_1987; @delhommeau_amelioration_1989; @delhommeau_seakeeping_1993].
@@ -43,16 +43,16 @@ Since its open-source release, it has been used in numerous applications, in par
 
 However, the capabilities of the code are limited in comparison with its commercial counterparts.
 Developing new features is difficult since the core of the implementation dating back from the 1970's is poorly readable and the documentation is scarce.
-The goal of the present work is thus the modernization of this code in order to have a open source linear potential flow solver that is easier to maintain and develop.
+The goal of the present work is the modernization of this code in order to have an open-source linear potential flow solver that is easier to maintain and develop.
 
 # The `capytaine` Python package
 
-The present work is a complete modernization of the code Nemoh, including a refactoring of the core routines and a Python user interface. 
+The present work is a complete modernization of the Nemoh code, including a refactoring of the core routines and a Python user interface. 
 
 The core routines have been kept in Fortran, although a more modern coding style has been used.
-Dead code has been removed, no global variable is used anymore, more meaningful names have been used for functions and variables and a lot of comments have been added.
+Dead code has been removed, no global variables are used anymore, more meaningful names have been used for functions and variables and a lot of comments have been added.
 
-This new version is meant to be used within the Python scientific ecosystem.
+This new version is meant to be used within the scientific Python ecosystem.
 The integration of the Fortran core with Python has been done with F2PY [@f2py].
 The naive linear solver included in Nemoh has been replaced by a call to the state-of-the-art solvers from the Numpy and Scipy libraries [@numpy; @scipy].
 The Intel MKL library can thus be used to solve the linear system in parallel with OpenMP.
@@ -60,11 +60,11 @@ The rest of the code is the independent computation of the coefficients of a mat
 
 The code has also been integrated with other Python libraries.
 Reading, writing and transforming meshes is done by the integration of the `meshmagick` [@meshmagick] library.
-The default output format is a `xarray` [@hoyer_xarray_2017] data set that can be saved in the standard NetCDF format.
+The default output format is a `xarray.Dataset` [@hoyer_xarray_2017] that can be saved in the standard NetCDF format.
[…]
