### 2019-03-19 #1335 @leouieda
:wave: Hi @mancellin I made a few editorial suggestions for the JOSS paper (mancellin/capytaine#6) and will start looking for reviewers now. Please let me know if you have any questions or suggestions for reviewers.
