### 2019-03-20 #1335 @leouieda
@mancellin please add these missing DOIs to your BibTex:

* https://doi.org/10.1115/OMAE2018-77924
* https://doi.org/10.1016/j.apor.2018.10.007
