### 2019-03-20 d8f358f
Add DOI in paper.bib and tidy the file.
