+++ paper/paper.bib
@@ -1,7 +1,7 @@
 
 @Article{	  alouges_fem_2018,
   author	= {Alouges, F. and Aussal, M.},
-  title		= {{FEM} and {BEM} simulations with the Gypsilab framework},
+  title		= {{FEM} and {BEM} simulations with the {Gypsilab} framework},
   journal	= {SMAI-Journal of computational mathematics},
   year		= {2018},
   volume	= {4},
@@ -11,27 +11,22 @@
 
 @Article{	  ancellin_hierarchical_2019,
   author	= {Ancellin, M. and Dias, F.},
-  title		= {Hierarchical Toeplitz matrices for the linear potential
-		  flow simulation of the floating bodies with local
-		  regularities},
+  title		= {Hierarchical {Toeplitz} matrices for the linear potential flow simulation of wave energy converters with local regularities},
   journal	= {in preparation},
   year		= {2019}
 }
 
 @InProceedings{	  ancellin_using_2018,
   author	= {Ancellin, M. and Dias, F.},
-  booktitle	= {{ASME 2018 37th International Conference on Ocean,
-		  Offshore and Arctic Engineering (OMAE2018)}},
-  title		= {Using the floating body symmetries to speed up the
-		  numerical computation of hydrodynamics coefficients with
-		  {N}emoh.},
+  booktitle	= {{ASME 2018 37th International Conference on Ocean, Offshore and Arctic Engineering (OMAE2018)}},
+  title		= {Using the floating body symmetries to speed up the numerical computation of hydrodynamics coefficients with {Nemoh}},
   year		= {2018},
-  pages		= {V009T13A031}
+  pages		= {V009T13A031},
+  doi       = {10.1115/OMAE2018-77924}
 }
 
 @Book{		  babarit_ocean_2017,
-  title		= {Ocean wave energy conversion: resource, technologies and
-		  performance},
+  title		= {Ocean wave energy conversion: resource, technologies and performance},
   author	= {Babarit, A.},
   year		= {2017},
   publisher	= {Elsevier}
@@ -63,18 +58,15 @@
 
 @InProceedings{	  delhommeau_seakeeping_1993,
   author	= {Delhommeau, G.},
-  title		= {The seakeeping codes Aquadyn and Aquaplus},
-  booktitle	= {19th WEGEMT School, Numerical Simulation of Hydrodynamics:
-		  Ships and Offshore Structures},
+  title		= {The seakeeping codes {Aquadyn} and {Aquaplus}},
+  booktitle	= {19th {WEGEMT} School, Numerical Simulation of Hydrodynamics: Ships and Offshore Structures},
   year		= {1993},
-  eventtitle	= {19th WEGEMT School, Numerical Simulation of Hydrodynamics:
-		  Ships and Offshore Structures}
+  eventtitle	= {19th {WEGEMT} School, Numerical Simulation of Hydrodynamics: Ships and Offshore Structures}
 }
 
 @Article{	  f2py,
   author	= {Peterson, P.},
-  title		= {{F2PY}: a Tool for Connecting Fortran and Python
-		  Programs},
+  title		= {{F2PY}: a Tool for Connecting Fortran and Python Programs},
   journal	= {Int. J. Comput. Sci. Eng.},
   volume	= {4},
   number	= {4},
@@ -83,22 +75,9 @@
   doi		= {10.1504/IJCSE.2009.029165}
 }
 
-@Article{	  fabregasflavia_numerical_2018,
-  author	= {F{\`a}bregas Flavi{\`a}, F and McNatt, C and Rong{\`e}re,
-		  F and Babarit, A and Cl{\'e}ment, AH},
[…]
