### 2019-03-20 #1341 @whedon
[REVIEW]: Capytaine: a Python-based linear potential flow solver
**Submitting author:** @mancellin (<a href="http://orcid.org/0000-0002-0316-3230">Matthieu Ancellin</a>)
**Repository:** <a href="https://github.com/mancellin/capytaine" target ="_blank">https://github.com/mancellin/capytaine</a>
**Version:** v1.0.1
