### 2019-03-21 7e46840
Fix compilation arguments.
