### 2019-03-22 #1341 @mikaem
@leouieda This is a strong submission and I'm ready to sign off as it is. My only real issue, that I don't consider a blocker, is that the authors should put a little more effort into the installation. I think they would benefit greatly from using both conda-forge and pypi, and from making this code available on mac and windows, not just linux. I could build it myself on mac using conda-forge compilers, using instructions from [here](https://conda-forge.org/docs/maintainer/knowledge_base.html).
