### 2019-03-22 93a279f
Fix minor issues in the docs.
Cf. Github issues #8 and #10.
