### 2019-03-22 50c7ede
Move some docstrings from __init__ to top of the class.
Cf. Github issue #10.
