### 2019-03-22 #1341 @leouieda
@mikaem thank you for your comments. Would you mind opening an issue on the software repository to address these issues? I agree that it's a non-blocker if you managed to get installed but our goal is to help improve the software through the review.

@mancellin would you like to try to fix these issues while we wait for @harpolea? I agree with @mikaem that using conda-forge is a better choice than using your own conda channel. I have a lot of experience with it and help set things up as well. I also agree that uploading to PyPI is a good thing as conda-forge often uses PyPI to get the source code for building packages.
