### 2019-03-25 mancellin/capytaine#11 @mikaem
Consider making capytaine available on conda-forge and PyPI
With conda-forge it should be much easier to maintain installable binaries for both osx, windows and linux. To help kickstart adding a recipe I have modified the meta.yaml you already have under the `conda` folder and, with the following sections, I could at least build locally on my mac:

    build:
      script: "{{ PYTHON }} -m pip install . --no-deps -vv"

    requirements:
      build:
        - {{ compiler('fortran') }}
        - {{ compiler('c') }}
        - llvm-openmp  # [osx]
      host:
        - python
        - pip
        - setuptools
        - numpy
        - llvm-openmp  # [osx]
      run:
        - python
        - attrs
        - {{ pin_compatible('numpy') }}
        - pandas
        - scipy
        - vtk
        - xarray
        - llvm-openmp  # [osx]

Also, see [guidelines for adding a recipe to conda-forge](https://conda-forge.org/#contribute) and a helpful [documentation](https://conda-forge.org/docs/index.html). Feel free to ping me if there are any issues with the pull request for the staged-recipes.

For uploading to [PyPI](https://pypi.org), please see [this link](https://packaging.python.org/tutorials/packaging-projects/).
