### 2019-03-25 #1341 @mikaem
@leouieda just opened an [issue](https://github.com/mancellin/capytaine/issues/11) with a suggestion for recipe. @mancellin needs to take it from there.
