### 2019-03-25 #1341 @mancellin
Yes, I agree that the distribution of the package should be improved.

I had not considered conda-forge, I will try it. Thank you @mikaem for the links and the recipe!
