### 2019-03-25 faf24f1
Fix license notation.
