+++ capytaine/__init__.py
@@ -5,7 +5,7 @@
 
 __author__ = 'Matthieu Ancellin'
 __version__ = '1.0'
-__license__ = 'GPLv3'
+__license__ = 'GPL-3.0'
 
 from capytaine.meshes.geometry import Axis, Plane, xOz_Plane, yOz_Plane, xOy_Plane
 from capytaine.meshes.meshes import Mesh
