### 2019-03-26 #1341 @mancellin
Mac and Linux packages are now available on conda-forge.
I still have some issues building a Windows package. I'll try to keep myself informed on the conda(-forge) Fortran toolchain for Windows.

It is also on PyPI, although only as a source package.
