### 2019-03-26 dba2b42
Update description in setup.py.
