### 2019-03-27 9a99e4b
Try to improve Windows building script.
