### 2019-03-28 8c19ca5
Add command to upload on PyPI in Makefile.
