### 2019-03-29 #1341 @leouieda
@mancellin thank you for keeping us updated :+1:

:wave: @harpolea just checking in: I see that there are some checklist items remaining for you. Please let me know when you have a chance to go through those remaining items.
