### 2019-04-01 #1341 @harpolea
Sorry again for the delay - I've finished up with the checks I had intended to do and recommend this for acceptance @leouieda

This is a really nice package - the documentation in particular is very thorough and was great for a new user like myself!
