### 2019-04-04 mancellin/capytaine#11 @mancellin
Hi @mikaem
Would you mind testing quickly that the conda-forge version works fine on your mac ?
Then, I suggest we close the issue.
