### 2019-04-08 #1341 @leouieda
@mancellin I have given the paper a final check and I'm happy to accept it for publication in JOSS :tada: :tada:

There a few steps that we need to take before publication. Please make a new release with any changes that can about after submission (if required) and archive a copy of the source code on Zenodo or figshare. Once that is done, please post the DOI of the archive here.
