### 2019-04-09 c95b07e
Fix typos in bibliography of JOSS paper.
