+++ paper/paper.bib
@@ -34,7 +34,7 @@
 
 @InProceedings{	  babarit_theoretical_2015,
   author	= {Babarit, A. and Delhommeau, G.},
-  title		= {Theoretical and numerical aspects of the open source BEM solver {NEMOH}},
+  title		= {Theoretical and numerical aspects of the open source {BEM} solver {NEMOH}},
   booktitle	= {{P}roceedings of the Eleventh {E}uropean {W}ave and {T}idal {E}nergy {C}onference ({EWTEC2015})},
   year		= {2015}
 }
@@ -51,7 +51,7 @@
 
 @InProceedings{delhommeau_amelioration_1989,
   author        = {Delhommeau, G.},
-  title         = {Amélioration des codes de calcul de diffraction-raditation au premier ordre},
+  title         = {Amélioration des codes de calcul de diffraction-radiation au premier ordre},
   booktitle     = {Actes des 2émes Journées de l'Hydrodynamique},
   year          = {1989},
 }
