### 2019-04-09 #1341 @mancellin
Thank you @leouieda, @harpolea and @mikaem!

I've read the paper once more and found a small typo in the bibliography. I have pushed a correction.

The DOI of the latest archive is 10.5281/zenodo.2633744
