### 2019-04-09 #1341 @leouieda
:wave: @openjournals/joss-eics this submission is ready to be accepted
