### 2019-04-10 #1341 @labarba
Congratulations, @mancellin, your JOSS paper is published!

Many thanks to our editor: @leouieda, and the reviewers: @mikaem, @harpolea — your contribution is greatly appreciated 🙏
